+++ drivers/ide.h
@@ -157,8 +157,8 @@
 typedef struct ide_part_ {
     uint8_t flags;
     uint8_t type;
-    uint32_t lba_start;
-    uint32_t lba_end;
+    uint64_t lba_start;
+    uint64_t lba_end;
 } ide_part_t;
 
 // Physical Region Descriptor
